infra: add global image loader
