+++ next.config.mjs
@@ -1,0 +1,22 @@
+import { setupDevPlatform } from "@cloudflare/next-on-pages/next-dev";
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  reactStrictMode: true,
+  images: {
+    loader: "custom",
+    loaderFile: "./imageLoader.ts",
+    remotePatterns: [
+      {
+        protocol: "https",
+        hostname: "**",
+      },
+    ],
+  },
+};
+
+if (process.env.NODE_ENV === "development") {
+  await setupDevPlatform();
+}
+
+export default nextConfig;
